modules/hw-function/streamlit/hello-world/requirements.txt
streamlit>=1.28.0
cognite-sdk
pyodide-http
